repost-bot/requirements.txt
python-telegram-bot>=21.0
vk-api>=11.9.9
requests>=2.31.0
python-dotenv>=1.0.0
